+++ src/main/java/life/weiwang/community/controller/QuestionController.java
@@ -15,7 +15,7 @@ public class QuestionController {
     QuestionService questionService;
 
     @GetMapping("/question/{id}")
-    public String question(@PathVariable(value = "id") Integer id,
+    public String question(@PathVariable(value = "id") Long id,
                            Model model) {
 
         QuestionDTO questionDTO = questionService.findById(id);
